add Voss algorithm for pink noise generation This is the first algorithm I used for the generation of pink noise.
The corresponding power spectrum did not looked that nice but it is a decent starting point since the output is definitely not white.

In addition, the production of the power spectrum is an art itself. I did not delved into that one up to now and just used the default package of R called `spectrum`. But the `psd` package looks promissing.